fix: refactor code structure and renaming
