+++ app/utils/__init__.py
@@ -1,3 +1,3 @@
-from .save_file import save_file_healthy, save_file_problem, save_file_as_json, is_cloudflare
+from .writer import save_file_healthy, save_file_problem, save_file_as_json, is_cloudflare
 from .output import sign, show_quiet, show_output, get_banner, print_legend
 from .summary import ReconStats
